Revert "All changed to float"
This reverts commit 49b2c6e1e87a3e044ef65488558d431f95d0a717.

+++ src/Solver/FFTPeriodicPoissonSolver.h
@@ -29,7 +29,7 @@
 namespace ippl {
 
     template <typename Tlhs, typename Trhs, unsigned Dim,
-              class M=UniformCartesian<float, Dim>,
+              class M=UniformCartesian<double, Dim>,
               class C=typename M::DefaultCentering>
     class FFTPeriodicPoissonSolver : public Electrostatics<Tlhs, Trhs, Dim, M, C>
     {
